nittiva-backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2
psycopg2-binary==2.9.9
drf-spectacular==0.27.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
google-auth==2.23.4
requests==2.31.0